% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_h5ad.R
\name{read_h5ad}
\alias{read_h5ad}
\title{Read data from h5ad file}
\usage{
read_h5ad(file = NULL, conda = NULL)
}
\arguments{
\item{file}{A file name or a path to h5ad file. Default = NULL.}

\item{conda}{a name of conda environment to use. Default = NULL.}
}
\value{
A convert object of AnnData
}
\description{
read data from h5ad file or the AnnData format.
}
\details{
This function requires Python AnnData.
}
\examples{

# Use the provided h5ad file with the package as example

\dontrun{
myfile <- system.file("extdata", "example.h5ad", package="episcanpy2r")
adata <- read_h5ad(myfile)
}
}
